+++ apps/ledger-live-desktop/src/renderer/actions/settings.js
@@ -6,7 +6,7 @@ import { useTranslation } from "react-i18next";
 import type { DeviceModelId } from "@ledgerhq/devices";
 import type { PortfolioRange } from "@ledgerhq/live-common/portfolio/v2/types";
 import type { Currency } from "@ledgerhq/types-cryptoassets";
-import type { DeviceModelInfo } from "@ledgerhq/types-live";
+import type { DeviceModelInfo, FeatureId, Feature } from "@ledgerhq/types-live";
 import { setEnvOnAllThreads } from "~/helpers/env";
 import type { SettingsState as Settings } from "~/renderer/reducers/settings";
 import {
@@ -242,4 +242,16 @@ export const removeStarredMarketCoins = (payload: string) => ({
 export const toggleStarredMarketCoins = (payload: string) => ({
   type: "TOGGLE_STARRED_MARKET_COINS",
   payload,
+});
+
+export const setOverriddenFeatureFlag = (key: FeatureId, value: Feature | undefined) => ({
+  type: "SET_OVERRIDDEN_FEATURE_FLAG",
+  payload: { key, value },
+});
+
+export const setOverriddenFeatureFlags = (overriddenFeatureFlags: {
+  [key: FeatureId]: Feature,
+}) => ({
+  type: "SET_OVERRIDDEN_FEATURE_FLAGS",
+  payload: { overriddenFeatureFlags },
 });