+++ apps/ledger-live-desktop/src/renderer/components/Onboarding/Screens/Tutorial/index.tsx
@@ -6,7 +6,6 @@ import {
   Icons,
   ProgressBar,
   Drawer,
-  Popin,
   InfiniteLoader,
 } from "@ledgerhq/react-ui";
 import React, { useCallback, useEffect, useMemo, useState } from "react";
@@ -16,6 +15,7 @@ import { Switch, Route, useHistory, useLocation } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import { useSelector } from "react-redux";
+import { Device } from "@ledgerhq/types-devices";
 import { languageSelector } from "~/renderer/reducers/settings";
 import { ImportYourRecoveryPhrase } from "~/renderer/components/Onboarding/Screens/Tutorial/screens/ImportYourRecoveryPhrase";
 import { DeviceHowTo } from "~/renderer/components/Onboarding/Screens/Tutorial/screens/DeviceHowTo";
@@ -39,7 +39,6 @@ import { QuizFailure } from "~/renderer/components/Onboarding/Screens/Tutorial/s
 import { QuizSuccess } from "~/renderer/components/Onboarding/Screens/Tutorial/screens/QuizSuccess";
 import RecoveryWarning from "../../Help/RecoveryWarning";
 import { QuizzPopin } from "~/renderer/modals/OnboardingQuizz/OnboardingQuizzModal";
-import { getPostOnboardingActionsForDevice } from "~/renderer/components/PostOnboardingHub/logic";
 import { useStartPostOnboardingCallback } from "@ledgerhq/live-common/postOnboarding/hooks/index";
 
 import { UseCase } from "../../index";
@@ -229,7 +228,7 @@ export default function Tutorial({ useCase }: Props) {
   const [userUnderstandConsequences, setUserUnderstandConsequences] = useState(false);
   const [userChosePinCodeHimself, setUserChosePinCodeHimself] = useState(false);
 
-  const [connectedDevice, setConnectedDevice] = useState(null);
+  const [connectedDevice, setConnectedDevice] = useState<Device | null>(null);
 
   const [onboardingDone, setOnboardingDone] = useState(false);
   const handleStartPostOnboarding = useStartPostOnboardingCallback();
@@ -488,14 +487,12 @@ export default function Tutorial({ useCase }: Props) {
         canContinue: !!connectedDevice,
         next: () => {
           setOnboardingDone(true);
-          handleStartPostOnboarding(connectedDevice.modelId, true, () => history.push("/"));
         },
         previous: () => history.push(`${path}/${ScreenId.pairMyNano}`),
       },
     ],
     [
       connectedDevice,
-      handleStartPostOnboarding,
       history,
       path,
       useCase,
@@ -514,14 +511,15 @@ export default function Tutorial({ useCase }: Props) {
        * block is executed), on the following commit we can call
        * history.push("/").
        */
-      const timeout = setTimeout(() => {
-        if (history.location.pathname !== "/") history.push("/");
+      const timeout: ReturnType<typeof setTimeout> = setTimeout(() => {
+        if (history.location.pathname !== "/")
+          handleStartPostOnboarding(connectedDevice.modelId, true, () => history.push("/"));
       }, 0);
       return () => {
         clearTimeout(timeout);
       };
     }
-  }, [history, onboardingDone]);
+  }, [connectedDevice?.modelId, handleStartPostOnboarding, history, onboardingDone]);
 
   const steps = useMemo(() => {
     const stepList = [