+++ libs/ledger-live-common/src/wallet-api/types.ts
@@ -1,0 +1,64 @@
+import type { SignedOperation } from "@ledgerhq/types-live";
+import type {
+  CryptoCurrency,
+  TokenCurrency,
+} from "@ledgerhq/types-cryptoassets";
+
+export type {
+  Families as WalletAPIFamilies,
+  Account as WalletAPIAccount,
+  Currency as WalletAPICurrency,
+  Transaction as WalletAPITransaction,
+  CryptoCurrency as WalletAPICryptoCurrency,
+  ERC20TokenCurrency as WalletAPIERC20TokenCurrency,
+} from "@ledgerhq/wallet-api-core";
+
+export {
+  CurrencyType as WalletAPICurrencyType,
+  TokenStandard as WalletAPITokenStandard,
+} from "@ledgerhq/wallet-api-core";
+
+export type TranslatableString = {
+  en: string;
+  [locale: string]: string;
+};
+
+export type AppPlatform =
+  | "desktop" // == windows || mac || linux
+  | "mobile" // == android || ios
+  | "all";
+
+export type AppBranch = "stable" | "experimental" | "soon" | "debug";
+
+export type AppPermission = {
+  method: string;
+  params?: any;
+};
+
+// TODO update to the new manifest types from wallet-api when released
+export type AppManifest = {
+  id: string;
+  private?: boolean;
+  name: string;
+  url: string | URL;
+  homepageUrl: string;
+  supportUrl?: string;
+  icon?: string | null;
+  platform: AppPlatform;
+  apiVersion: string;
+  manifestVersion: string;
+  branch: AppBranch;
+  params?: string[];
+  categories: string[];
+  currencies: string[] | "*";
+  content: {
+    shortDescription: TranslatableString;
+    description: TranslatableString;
+  };
+  permissions: AppPermission[];
+  domains: string[];
+};
+
+export type WalletAPISignedTransaction = SignedOperation;
+
+export type WalletAPISupportedCurrency = CryptoCurrency | TokenCurrency;