+++ libs/ledgerjs/packages/devices/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/devices
 
+## 7.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`f1c15446da`](https://github.com/LedgerHQ/ledger-live/commit/f1c15446dabef05bb91dada8d8f53f9bc6474ba5)]:
+  - @ledgerhq/errors@6.12.2
+
+## 7.0.6-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`f1c15446da`](https://github.com/LedgerHQ/ledger-live/commit/f1c15446dabef05bb91dada8d8f53f9bc6474ba5)]:
+  - @ledgerhq/errors@6.12.2-next.0
+
 ## 7.0.5
 
 ### Patch Changes